renamed files to match project convention
