+++ src/lead_registration/lead_registration.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { LeadRegistrationService } from './lead_registration.service';
 import { LeadRegistrationController } from './lead_registration.controller';
-import { Registration, RegistrationSchema } from 'src/shared/schema/lead_registration.schema';
+import { Registration, RegistrationSchema } from 'src/shared/schema/lead-registration.schema';
 import { UsersModule } from '../users/users.module'; // Import the UsersModule
 import { User, UserSchema } from 'src/shared/schema';
-import { TempLeadRegistration, TempLeadRegistrationSchema } from 'src/shared/schema/temp_lead_registration.schema';
+import { TempLeadRegistration, TempLeadRegistrationSchema } from 'src/shared/schema/temp-lead_registration.schema';
 
 
 @Module({
