+++ src/shared/schema/lead-registration.schema.ts
