+++ src/shared/schema/temp-lead_registration.schema.ts
